pharmacy_service/requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
psycopg2-binary>=2.9.10
requests>=2.32.0
PyJWT>=2.9.0
mysqlclient>=2.2.0
